fix: fix beq command encoding
fixed a bug in the assembly process for branch commands. First, the check for a stored was moved prior to validation of actual numeric inputs. turns out labels can be <=3 in length. crazy. second, fixed issue with invalid PC jumping by adding in the correct hex-radix to a string conversion. changed the addition overflow handling to a mod vs subtract due to paranoia.